requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.1
urllib3==1.26.13
